Extract DeepKeyValue in its own module
