+++ getaround_utils/lib/getaround_utils.rb
@@ -4,4 +4,5 @@ module GetaroundUtils
   autoload :LogFormatters, 'getaround_utils/log_formatters'
   autoload :Mixins, 'getaround_utils/mixins'
   autoload :Patches, 'getaround_utils/patches'
+  autoload :Utils, 'getaround_utils/utils'
 end
