+++ getaround_utils/lib/getaround_utils/utils.rb
@@ -1,0 +1,3 @@
+module GetaroundUtils::Utils
+  autoload :DeepKeyValueSerializer, 'getaround_utils/utils/deep_key_value_serializer'
+end
